requirements.txt
ckuehl-upsidedown
discord-py
parsedatetime
